+++ src/test/java/racingcar/domain/CarFactoryTest.java
@@ -1,8 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Car;
+import racingcar.domain.CarFactory;
 
 import java.util.List;
